+++ src/main/java/by/dz/fruits_in_choco/fruits_in_choco/controller/UserController.java
@@ -31,7 +31,7 @@ public class UserController {
 
     @PreAuthorize("hasAuthority('ADMIN')")
     @GetMapping("/admin/users/{id}")
-    public ResponseEntity<?> getUserById(@PathVariable Long id) {
+    public ResponseEntity<?> getUserById(@PathVariable Short id) {
         return ResponseEntity.ok(userMapper.getUserById(id));
     }
 }
