+++ src/main/java/by/dz/fruits_in_choco/fruits_in_choco/mapper/CategoryMapper.java
@@ -28,7 +28,7 @@ public class CategoryMapper {
                 .collect(Collectors.toList());
     }
 
-    public CategoryResponse getCategoryById(Long id) {
+    public CategoryResponse getCategoryById(Short id) {
         return mapToResponseDTO(service.getCategoryById(id));
     }
