+++ src/main/java/by/dz/fruits_in_choco/fruits_in_choco/mapper/UserMapper.java
@@ -47,7 +47,7 @@ public class UserMapper {
         return mapToResponseDto(userService.getUserByEmail(email));
     }
 
-    public UserResponse getUserById(Long id) {
+    public UserResponse getUserById(Short id) {
         return mapToResponseDto(userService.getUserById(id));
     }
