+++ src/main/java/by/dz/fruits_in_choco/fruits_in_choco/repository/CakeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CakeRepository extends JpaRepository<Cake, Long> {
+public interface CakeRepository extends JpaRepository<Cake, Short> {
 }
