+++ src/main/java/by/dz/fruits_in_choco/fruits_in_choco/repository/CartItemRepository.java
@@ -8,5 +8,5 @@ import javax.transaction.Transactional;
 
 @Repository
 @Transactional
-public interface CartItemRepository extends JpaRepository<CartItem, Long> {
+public interface CartItemRepository extends JpaRepository<CartItem, Short> {
 }
