+++ src/main/java/by/dz/fruits_in_choco/fruits_in_choco/repository/IngredientRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface IngredientRepository extends JpaRepository<Ingredient, Long> {
+public interface IngredientRepository extends JpaRepository<Ingredient, Short> {
     List<Ingredient> findByStatus(CakeIngredientStatus status);
 }
