+++ src/main/java/by/dz/fruits_in_choco/fruits_in_choco/service/CategoryService.java
@@ -6,10 +6,10 @@ import by.dz.fruits_in_choco.fruits_in_choco.entity.category.Category;
 import java.util.List;
 
 public interface CategoryService {
-    Category getCategoryById(Long id);
+    Category getCategoryById(Short id);
     List<Category> getCategories(int page, int size, String direction, String sortBy);
     Category saveCategory(CategoryRequest request);
-    Category updateCategory(Category category, Long id);
-    void deleteCategoryById(Long id);
+    Category updateCategory(Category category, Short id);
+    void deleteCategoryById(Short id);
     Category getCategoryByName(String name);
 }
