+++ src/main/java/by/dz/fruits_in_choco/fruits_in_choco/service/impl/UserServiceImpl.java
@@ -30,7 +30,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User getUserById(Long id) {
+    public User getUserById(Short id) {
         return userRepository.findById(id).get();
     }
